+++ lib/config/options.rb
@@ -1,10 +1,10 @@
 require 'ostruct'
-require 'config/validation/validate' if RUBY_VERSION >= '2.1'
+require 'config/validation/validate'
 
 module Config
   class Options < OpenStruct
     include Enumerable
-    include Validation::Validate if RUBY_VERSION >= '2.1'
+    include Validation::Validate
 
     def keys
       marshal_dump.keys
@@ -16,7 +16,7 @@ module Config
 
     def add_source!(source)
       # handle yaml file paths
-      source = (Sources::YAMLSource.new(source)) if source.is_a?(String)
+      source = (Sources::YAMLSource.new(source)) if source.is_a?(String) || source.is_a?(Pathname)
       source = (Sources::HashSource.new(source)) if source.is_a?(Hash)
 
       @config_sources ||= []
@@ -24,45 +24,12 @@ module Config
     end
 
     def prepend_source!(source)
-      source = (Sources::YAMLSource.new(source)) if source.is_a?(String)
+      source = (Sources::YAMLSource.new(source)) if source.is_a?(String) || source.is_a?(Pathname)
       source = (Sources::HashSource.new(source)) if source.is_a?(Hash)
 
       @config_sources ||= []
       @config_sources.unshift(source)
     end
-
-    def reload_env!
-      return self if ENV.nil? || ENV.empty?
-
-      hash = Hash.new
-
-      ENV.each do |variable, value|
-        keys = variable.to_s.split(Config.env_separator)
-
-        next if keys.shift != (Config.env_prefix || Config.const_name)
-
-        keys.map! { |key|
-          case Config.env_converter
-            when :downcase then
-              key.downcase.to_sym
-            when nil then
-              key.to_sym
-            else
-              raise "Invalid ENV variables name converter: #{Config.env_converter}"
-          end
-        }
-
-        leaf = keys[0...-1].inject(hash) { |h, key|
-          h[key] ||= {}
-        }
-
-        leaf[keys.last] = Config.env_parse_values ? __value(value) : value
-      end
-
-      merge!(hash)
-    end
-
-    alias :load_env! :reload_env!
 
     # look through all our sources and rebuild the configuration
     def reload!
@@ -73,19 +40,22 @@ module Config
         if conf.empty?
           conf = source_conf
         else
-          DeepMerge.deep_merge!(source_conf,
+          DeepMerge.deep_merge!(
+                                source_conf,
                                 conf,
                                 preserve_unmergeables: false,
                                 knockout_prefix:       Config.knockout_prefix,
-                                overwrite_arrays:      Config.overwrite_arrays)
+                                overwrite_arrays:      Config.overwrite_arrays,
+                                merge_nil_values:      Config.merge_nil_values,
+                                merge_hash_arrays:     Config.merge_hash_arrays
+                               )
         end
       end
 
       # swap out the contents of the OStruct with a hash (need to recursively convert)
       marshal_load(__convert(conf).marshal_dump)
 
-      reload_env! if Config.use_env
-      validate! if RUBY_VERSION >= '2.1'
+      validate!
 
       self
     end
@@ -111,6 +81,8 @@ module Config
       result
     end
 
+    alias :to_h :to_hash
+
     def each(*args, &block)
       marshal_dump.each(*args, &block)
     end
@@ -120,18 +92,27 @@ module Config
       to_hash.to_json(*args)
     end
 
+    def as_json(options = nil)
+      to_hash.as_json(options)
+    end
+
     def merge!(hash)
       current = to_hash
-      DeepMerge.deep_merge!(hash.dup,
+      DeepMerge.deep_merge!(
+                            hash.dup,
                             current,
                             preserve_unmergeables: false,
-                            overwrite_arrays:      Config.overwrite_arrays)
+                            knockout_prefix:       Config.knockout_prefix,
+                            overwrite_arrays:      Config.overwrite_arrays,
+                            merge_nil_values:      Config.merge_nil_values,
+                            merge_hash_arrays:     Config.merge_hash_arrays
+                           )
       marshal_load(__convert(current).marshal_dump)
       self
     end
 
     # Some keywords that don't play nicely with OpenStruct
-    SETTINGS_RESERVED_NAMES = %w{select collect test count}
+    SETTINGS_RESERVED_NAMES = %w[select collect test count zip min max exit!].freeze
 
     # An alternative mechanism for property access.
     # This let's you do foo['bar'] along with foo.bar.
@@ -148,6 +129,25 @@ module Config
       define_method name do
         self[name]
       end
+    end
+
+    def key?(key)
+      table.key?(key)
+    end
+
+    def has_key?(key)
+      table.has_key?(key)
+    end
+
+    def method_missing(method_name, *args)
+      if Config.fail_on_missing && method_name !~ /.*(?==\z)/m
+        raise KeyError, "key not found: #{method_name.inspect}" unless key?(method_name)
+      end
+      super
+    end
+
+    def respond_to_missing?(*args)
+      super
     end
 
     protected
@@ -170,7 +170,6 @@ module Config
 
       h.each do |k, v|
         k = k.to_s if !k.respond_to?(:to_sym) && k.respond_to?(:to_s)
-        s.new_ostruct_member(k)
 
         if v.is_a?(Hash)
           v = v["type"] == "hash" ? v["contents"] : __convert(v)
@@ -178,14 +177,9 @@ module Config
           v = v.collect { |e| e.instance_of?(Hash) ? __convert(e) : e }
         end
 
-        s.send("#{k}=".to_sym, v)
+        s[k] = v
       end
       s
-    end
-
-    # Try to convert string to a correct type
-    def __value(v)
-      Integer(v) rescue Float(v) rescue v
     end
   end
 end