+++ spec/config_env_spec.rb
@@ -1,14 +1,13 @@
 require 'spec_helper'
 
-describe Config do
+describe Config::Options do
+  before :each do
+    Config.reset
+  end
 
   context 'when overriding settings via ENV variables is enabled' do
     let(:config) do
       Config.load_files "#{fixture_path}/settings.yml", "#{fixture_path}/multilevel.yml"
-    end
-
-    before :all do
-      Config.use_env = true
     end
 
     after :all do
@@ -18,10 +17,11 @@ describe Config do
     before :each do
       ENV.clear
 
-      Config.env_prefix       = nil
-      Config.env_separator    = '.'
-      Config.env_converter    = :downcase
-      Config.env_parse_values = true
+      Config.use_env              = true
+      Config.env_prefix           = nil
+      Config.env_separator        = '.'
+      Config.env_converter        = :downcase
+      Config.env_parse_values     = true
     end
 
     it 'should add new setting from ENV variable' do
@@ -47,36 +47,52 @@ describe Config do
       end
     end
 
-    context 'and parsing ENV variable names is enabled' do
-      it 'should recognize numbers and expose them as integers' do
-        ENV['Settings.new_var'] = '123'
+    context 'and parsing ENV variable values' do
+      context 'is enabled' do
+        it 'should recognize "false" and expose as Boolean' do
+          ENV['Settings.new_var'] = 'false'
 
-        expect(config.new_var).to eq(123)
-        expect(config.new_var.is_a? Integer).to eq(true)
+          expect(config.new_var).to eq(false)
+          expect(config.new_var.is_a? FalseClass).to eq(true)
+        end
+
+        it 'should recognize "true" and expose as Boolean' do
+          ENV['Settings.new_var'] = 'true'
+
+          expect(config.new_var).to eq(true)
+          expect(config.new_var.is_a? TrueClass).to eq(true)
+        end
+
+        it 'should recognize numbers and expose them as integers' do
+          ENV['Settings.new_var'] = '123'
+
+          expect(config.new_var).to eq(123)
+          expect(config.new_var.is_a? Integer).to eq(true)
+        end
+
+        it 'should recognize fixed point numbers and expose them as float' do
+          ENV['Settings.new_var'] = '1.9'
+
+          expect(config.new_var).to eq(1.9)
+          expect(config.new_var.is_a? Float).to eq(true)
+        end
+
+        it 'should leave strings intact' do
+          ENV['Settings.new_var'] = 'foobar'
+
+          expect(config.new_var).to eq('foobar')
+          expect(config.new_var.is_a? String).to eq(true)
+        end
       end
 
-      it 'should recognize fixed point numbers and expose them as float' do
-        ENV['Settings.new_var'] = '1.9'
+      context 'is disabled' do
+        it 'should not convert numbers to integers' do
+          ENV['Settings.new_var'] = '123'
 
-        expect(config.new_var).to eq(1.9)
-        expect(config.new_var.is_a? Float).to eq(true)
-      end
+          Config.env_parse_values = false
 
-      it 'should leave strings intact' do
-        ENV['Settings.new_var'] = 'foobar'
-
-        expect(config.new_var).to eq('foobar')
-        expect(config.new_var.is_a? String).to eq(true)
-      end
-    end
-
-    context 'and parsing ENV variable names is disabled' do
-      it 'should not convert numbers to integers' do
-        ENV['Settings.new_var'] = '123'
-
-        Config.env_parse_values = false
-
-        expect(config.new_var).to eq('123')
+          expect(config.new_var).to eq('123')
+        end
       end
     end
 
@@ -126,22 +142,65 @@ describe Config do
       end
     end
 
-    context 'and variable names conversion is enabled' do
-      it 'should downcase variable names when :downcase conversion enabled' do
-        ENV['Settings.NEW_VAR'] = 'value'
+    context 'and custom ENV variables prefix includes custom ENV variables separator' do
+      before :each do
+        Config.env_prefix = 'MY_CONFIG'
+        Config.env_separator = '_'
+      end
 
-        expect(config.new_var).to eq('value')
+      it 'should load environment variables which begin with the custom prefix' do
+        ENV['MY_CONFIG_KEY'] = 'value'
+
+        expect(config.key).to eq('value')
+      end
+
+      it 'should not load environment variables which begin with the default prefix' do
+        ENV['Settings_key'] = 'value'
+
+        expect(config.key).to eq(nil)
+      end
+
+      it 'should not load environment variables which partially begin with the custom prefix' do
+        ENV['MY_CONFIGS_KEY'] = 'value'
+
+        expect(config.key).to eq(nil)
+      end
+
+      it 'should recognize the custom separator' do
+        ENV['MY_CONFIG_KEY.WITH.DOT']           = 'value'
+        ENV['MY_CONFIG_WORLD_COUNTRIES_EUROPE'] = '0'
+
+        expect(config['key.with.dot']).to eq('value')
+        expect(config.world.countries.europe).to eq(0)
+      end
+
+      it 'should not recognize the default separator' do
+        ENV['MY_CONFIG.KEY'] = 'value'
+
+        expect(config.key).to eq(nil)
       end
     end
 
-    context 'and variable names conversion is disabled' do
-      it 'should not change variable names by default' do
-        ENV['Settings.NEW_VAR'] = 'value'
+    context 'and variable names conversion' do
+      context 'is enabled' do
+        it 'should downcase variable names when :downcase conversion enabled' do
+          ENV['Settings.NEW_VAR'] = 'value'
 
-        Config.env_converter = nil
+          Config.env_converter = :downcase
 
-        expect(config.new_var).to eq(nil)
-        expect(config.NEW_VAR).to eq('value')
+          expect(config.new_var).to eq('value')
+        end
+      end
+
+      context 'is disabled' do
+        it 'should not change variable names by default' do
+          ENV['Settings.NEW_VAR'] = 'value'
+
+          Config.env_converter = nil
+
+          expect(config.new_var).to eq(nil)
+          expect(config.NEW_VAR).to eq('value')
+        end
       end
     end
 
@@ -155,5 +214,23 @@ describe Config do
       expect(config.new_var).to eq('value')
     end
 
+    context 'and env variable names conflict with new namespaces' do
+      it 'should throw a descriptive error message' do
+        ENV['Settings.backend_database'] = 'development'
+        ENV['Settings.backend_database.user'] = 'postgres'
+
+        expected_message = 'Environment variable Settings.backend_database.user '\
+          'conflicts with variable Settings.backend_database'
+        expect { config }.to raise_error(RuntimeError, expected_message)
+      end
+    end
+
+    context 'and env variable names conflict with existing namespaces' do
+      it 'should allow overriding the namespace' do
+        ENV['Settings.databases'] = 'new databases'
+
+        expect(config.databases).to eq('new databases')
+      end
+    end
   end
 end