+++ spec/spec_helper.rb
@@ -1,9 +1,12 @@
 ENV['RAILS_ENV'] ||= 'test'
 
+puts "RUBY_ENGINE: #{RUBY_ENGINE}"
+puts "RUBY_VERSION: #{RUBY_VERSION}\n\n"
+
 ##
 # Code Climate
 #
-if ENV['TRAVIS']
+if ENV['GITHUB_ACTIONS'] && RUBY_ENGINE == 'ruby' && RUBY_VERSION.start_with?(ENV['COVERAGE_RUBY_VERSION'] || 'x')
   require 'simplecov'
   SimpleCov.start
 end
@@ -14,28 +17,52 @@ end
 Dir['./spec/support/**/*.rb'].each { |f| require f }
 
 ##
-# Load Rails dummy application based on gemfile name substituted by Appraisal
+# Detect Rails/Sinatra dummy application based on gemfile name substituted by Appraisal
 #
-if ENV['APPRAISAL_INITIALIZED'] || ENV['TRAVIS']
+if ENV['APPRAISAL_INITIALIZED'] || ENV['GITHUB_ACTIONS']
   app_name = Pathname.new(ENV['BUNDLE_GEMFILE']).basename.sub('.gemfile', '')
 else
-  app_name = 'rails_5'
+  /.*?(?<app_name>rails.*?)\.gemfile/ =~ Dir["gemfiles/rails*.gemfile"].sort.last
 end
 
-require File.expand_path("../../spec/app/#{app_name}/config/environment", __FILE__)
+##
+# Load dummy application and Rspec
+#
+app_framework = %w{rails sinatra}.find { |f| app_name.to_s.include?(f) }
 
-APP_RAKEFILE = File.expand_path("../../spec/app/#{app_name}/Rakefile", __FILE__)
+case app_framework
+when 'rails'
+  # Load Rails
+  require File.expand_path("../app/#{app_name}/config/environment", __FILE__)
+
+  APP_RAKEFILE = File.expand_path("../app/#{app_name}/Rakefile", __FILE__)
+
+  # Load Rspec
+  require 'rspec/rails'
+
+  # Configure
+  RSpec.configure do |config|
+    config.fixture_path = FixtureHelper::FIXTURE_PATH
+  end
+
+when 'sinatra'
+  # Load Sinatra
+  require File.expand_path("../app/#{app_name}/app", __FILE__)
+
+  # Load Rspec
+  require 'rspec'
+
+  # Configure
+  RSpec.configure do |config|
+    config.filter_run_excluding :rails
+    config.include FixtureHelper
+  end
+end
 
 ##
-# Load Rspec
+# Common Rspec configure
 #
-require 'rspec/rails'
-require 'pry'
-
-# Configure
 RSpec.configure do |config|
-  config.fixture_path = File.join(File.dirname(__FILE__), '/fixtures')
-
   # Turn the deprecation warnings into errors, giving you the full backtrace
   config.raise_errors_for_deprecations!
 
@@ -44,27 +71,29 @@ RSpec.configure do |config|
 
       # Extend Config module with ability to reset configuration to the default values
       def self.reset
-        self.const_name       = 'Settings'
-        self.use_env          = false
-        self.knockout_prefix  = nil
-        self.overwrite_arrays = true
-        self.schema           = nil if RUBY_VERSION >= '2.1'
-        class_variable_set(:@@_ran_once, false)
+        self.const_name           = 'Settings'
+        self.use_env              = false
+        self.knockout_prefix      = nil
+        self.overwrite_arrays     = true
+        self.schema               = nil
+        self.validation_contract  = nil
+        self.fail_on_missing      = false
+        instance_variable_set(:@_ran_once, false)
       end
     end
   end
 end
-
 
 ##
 # Print some debug info
 #
 puts
 puts "Gemfile: #{ENV['BUNDLE_GEMFILE']}"
-puts 'Rails version:'
+puts 'Version:'
 
 Gem.loaded_specs.each { |name, spec|
-  puts "\t#{name}-#{spec.version}" if %w{rails activesupport sqlite3 rspec-rails}.include?(name)
+  puts "\t#{name}-#{spec.version}" if %w{rails activerecord-jdbcsqlite3-adapter sqlite3 rspec-rails sinatra}.include?(name)
 }
+puts "\tpsych-#{Psych::VERSION}"
 
 puts