+++ spec/tasks/db_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe 'db:create' do
+describe 'db:create', :rails do
   include_context 'rake'
 
   it 'has access to Settings object and can read databases from settings.yml file' do